requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.2.1
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
pytz==2023.3
sqlparse==0.4.4
asgiref==3.6.0
black==23.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.5
Django==4.2.1
django-allauth==0.50.0
django-colorfield==0.7.1
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
filelock==3.6.0
idna==3.3
mypy-extensions==1.0.0
mysql-connector-python==8.0.29
oauthlib==3.2.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pipenv==2022.3.28
platformdirs==3.5.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2023.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.0
virtualenv-clone==0.5.7
